Update to dbinit.sql
Here are the proposed changes. My code is in SQL. Based on what you've offered so far, I think I have an idea of how to convert what needed to be changed, but I want to be sure before I do. I apologize if I pasted this in the incorrect file.