Update to dbinit.sql
According to pgAdmin III, the code executes properly--I had to learn tables need to be made in a proper order depending on the dependencies between tables. Speaking of dependencies, I also checked to make sure primary keys and foreign keys were assigned appropriately. Let me know if there are problems with any of this, otherwise.